Yolo5/requirements.txt
pyyaml
loguru
boto3
